services/lambda-label-successful-prs/requirements.txt
boto3==1.14.28
botocore==1.17.28
certifi==2020.6.20
chardet==3.0.4
Deprecated==1.2.10
docutils==0.15.2
idna==2.10
jenkinsapi==0.3.11
jmespath==0.10.0
PyGithub==1.51
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
wrapt==1.12.1
